hearo + chime: first wmake clean, boot screen renders, tests pass
Open Watcom V2 installed; first compile surfaced the expected handful
of issues. All fixed.

Build:
  - hearo Makefile: add -zc to CF16. With ~70 unlock rules, menu
    definitions, and status-output format strings, DGROUP overflowed
    64K by 10304 bytes. -zc places const data in the code segment.
  - hearo + chime Makefiles: split LIB_OBJS from main entry. Test
    programs link with LIB_OBJS only; HEARO.EXE / CHIME.EXE add the
    main-bearing object on top. Stops the 'redefinition of main_'
    linker warning when test programs link with the full object list.
  - hearo/src/detect/detect.c: drop the duplicate dos_get_date stub.
    platform/dos.c is the canonical definition; detect.c was redefining
    the symbol under HEARO_NOASM.
  - hearo/src/platform/probes.c: rephrase the leading comment block to
    avoid '*.c' patterns Watcom read as nested comment terminators.
  - hearo/src/unlock/unlock.c: drop unused chk_386plus.
  - hearo/src/math/bipartite.c: bipartite_sin had two bugs. First, the
    s32 sum truncated to s16 wrapped to negative when both A and B
    table entries were near the s16 max, producing 65448-magnitude
    error spikes near peaks. Second, the bp_sin_b residual table only
    encoded sin's quadrant shape, not the actual residual after the
    coarse step. Replaced with linear interpolation between adjacent
    A entries plus a clamp. TESTBIP now reports max err 3 (~13 bits)
    against 1024 reference samples; previous build was ~5.5 bits even
    after the clamp fix.
  - hearo/test/testquir.c: drop the s64 cast in the naive sum path.
    s64 is not declared at test scope and the overflow is the whole
    point of the comparison anyway.

HEARO_NOASM scope narrowed:
  - Was: gates every Watcom-ism (display VRAM, DOS calls, file system).
  - Now: gates only the inline-asm probes in src/platform/probes.c.
  - hearo/src/ui/screen.c, hearo/src/platform/dos.c,
    hearo/src/platform/timer.c, hearo/src/ui/browser.c, and
    chime/src/dos_clock.c lose their HEARO_NOASM stubs entirely. They
    always use the real Watcom path (B800:0000 VRAM, intdos, BIOS tick,
    _dos_findfirst, port 0x70/0x71 CMOS).
  - Without this, the boot screen rendered to an in-memory emu_buffer
    instead of real video memory, which is why the first capture
    attempt showed only the autoexec output. With this, the boot screen
    renders for real and reads back via DOSBox-X PrintWindow.

Screenshot toolkit:
  - screenshot.ps1: add -fastlaunch to the DOSBox-X args. Skips the
    BIOS logo and welcome banner that otherwise overlays the captured
    frame.
  - screenshot.ps1: default OutFile is now C:\Development\Screenshots\
    so screenshots land alongside cathode/discord/etc instead of in a
    sibling NetISA\Screenshots\ directory.
  - screenshot.conf: drop the diagnostic 'echo Starting HEARO' that
    crept in during debugging.
  - new test-cpu386.conf in hearo/scripts/ and chime/scripts/ for
    headless test runs (autoexec runs the test executables, redirects
    stdout to a file, exits).

Test results from DOSBox-X 386 conf:
  - TESTCORD: precision improves monotonically from 8 to 24 iterations
    as expected. max-err-sin goes from 475 at 8 iters to 5 at 16 iters
    (Q16.16 precision floor).
  - TESTBIP: max err 3 across 1024 samples (~13 bits accuracy).
  - TESTQUIR: naive sum overflows to -462758661, quire low word matches
    naive but the high word holds the overflow carry (4294967294).
    Quire is nonzero across the full 256-bit state.
  - TESTDET: detection produces a Workstation-tier profile (486DX +
    integrated FPU + VESA SVGA + 16 MB XMS), 24 of 67 features
    unlocked, FNV-1a fingerprint stable.
  - HEARO /VERSION and /UNLOCKS print correctly.
  - HEARO boot screen captured at /Screenshots/hearo_boot.png:
    CP437 box-drawing logo, hardware report with right-justified
    ENABLED in green, indented unlock list, "Your machine has unlocked
    24 features. Press any key to begin." closing line.
  - CHIME TESTTIME: 21 of 21 assertions pass (unix_ts roundtrip,
    HTTP Date parser, worldtime JSON parser, stub query).

+++ hearo/test/testquir.c
@@ -31,7 +31,7 @@ int main(void)
             a = -((s32)(rand() % 30000) + 1);
             b = (s32)(rand() % 30000) + 1;
         }
-        naive += (s32)((s64)a * (s64)b);
+        naive += a * b;  /* deliberate s32 overflow: that is what the quire fixes */
         quire_mac(&q, a, b);
     }
     {
